// backend/src/pet/enums/element.enum.ts
export enum Element {
  FEUER = 'feuer',
  WASSER = 'wasser',
  WIND = 'wind',
  ERDE = 'erde',
  BLITZ = 'blitz',
  METALL = 'metall',
  LICHT = 'licht',
  SCHATTEN = 'schatten',
  CHAOS = 'chaos',
  KRISTALL = 'kristall',
  NATUR = 'natur',
  GEIST = 'geist',
  FEE = 'fee',
  EIS = 'eis',
  GIFT = 'gift',
  KOSMOS = 'kosmos',
}